# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travana API Documentation",
        "description": "Complete API documentation for Travana - A travel and vacation rental platform. This API provides endpoints for managing listings, reservations, users, blog content, and more.",
        "contact": {
            "name": "Travana Support",
            "email": "support@travana.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.travana.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/blog-categories": {
            "get": {
                "tags": [
                    "Blog Categories"
                ],
                "summary": "Get all blog categories",
                "operationId": "getBlogCategories",
                "parameters": [
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Filter active only",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Return all (no pagination)",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of blog categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BlogCategory"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog Categories"
                ],
                "summary": "Create a new blog category",
                "operationId": "createBlogCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "slug": {
                                        "description": "Auto-generated if not provided",
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "icon": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "color": {
                                        "type": "string",
                                        "maxLength": 7,
                                        "example": "#3498db"
                                    },
                                    "meta_title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "meta_description": {
                                        "type": "string",
                                        "maxLength": 500
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "category": {
                                            "$ref": "#/components/schemas/BlogCategory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/blog-categories/{identifier}": {
            "get": {
                "tags": [
                    "Blog Categories"
                ],
                "summary": "Get a specific blog category",
                "operationId": "getBlogCategory",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Category ID or slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BlogCategory"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog-categories/{id}": {
            "put": {
                "tags": [
                    "Blog Categories"
                ],
                "summary": "Update a blog category",
                "operationId": "updateBlogCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "icon": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "meta_title": {
                                        "type": "string"
                                    },
                                    "meta_description": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "category": {
                                            "$ref": "#/components/schemas/BlogCategory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blog Categories"
                ],
                "summary": "Delete a blog category",
                "operationId": "deleteBlogCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cannot delete - has blogs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BlogCategory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Travel Tips"
                    },
                    "slug": {
                        "type": "string",
                        "example": "travel-tips"
                    },
                    "description": {
                        "type": "string",
                        "example": "Helpful travel advice",
                        "nullable": true
                    },
                    "icon": {
                        "type": "string",
                        "example": "fa-plane",
                        "nullable": true
                    },
                    "color": {
                        "type": "string",
                        "example": "#3498db",
                        "nullable": true
                    },
                    "meta_title": {
                        "type": "string",
                        "nullable": true
                    },
                    "meta_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "order": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "blogs_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "message": {
                        "type": "string",
                        "example": "Detailed error message"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field_name": [
                                "The field_name is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}